chore: extract TemplateSelector
