+++ src/core_modules/capture-core/components/TemplateSelector/TemplateSelectorChip.component.js
