+++ src/core_modules/capture-core/components/TemplateSelector/index.js
