+++ src/core_modules/capture-core/components/TemplateSelector/useTemplates.js
