+++ src/core_modules/capture-core/components/TemplateSelector/workingListsBase.types.js
